requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
celery==5.3.4
Django==4.2.5
djangorestframework==3.14.0
pytz==2023.3.post1
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
flake8-isort==6.0.0